🎨 Statistics 내 필드타입 래퍼 타입으로 변경

+++ src/main/java/com/nexters/keyme/domain/test/presentation/dto/response/SingleTestStatisticsResponse.java
@@ -15,7 +15,7 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class SingleTestStatisticsResponse {
-    private float averageRate;
-    private int solvedCount;
+    private Double averageRate;
+    private Long solvedCount;
     private List<QuestionStatisticResponse> questionsStatistics;
 }
